Replace native base button with custom component for school filter button
